@@ -57,6 +57,7 @@
     "@loopback/security": "^0.1.13",
     "@loopback/service-proxy": "^1.3.17",
     "bcryptjs": "^2.4.3",
+    "jsonwebtoken": "^8.5.1",
     "loopback-connector-postgresql": "^3.8.3"
   },
   "devDependencies": {
@@ -72,6 +73,6 @@
     "eslint-plugin-eslint-plugin": "^2.2.1",
     "eslint-plugin-mocha": "^6.2.2",
     "source-map-support": "^0.5.16",
-    "typescript": "~3.7.5"
+    "typescript": "~3.8.2"
   }
 }
