@@ -1,5 +1,18 @@
 import {Entity, model, property} from '@loopback/repository';
 
+// @model({
+//   settings: {
+//     foreignKeys: {
+//       fk_credential_userId: {
+//         name: 'fk_credential_userId',
+//         entity: 'User',
+//         entityKey: 'email',
+//         foreignKey: 'userId',
+//       },
+//     },
+//   },
+// })
+
 @model()
 export class UserCredentials extends Entity {
   @property({
@@ -21,7 +34,6 @@ export class UserCredentials extends Entity {
   })
   password: string;
 
-
   constructor(data?: Partial<UserCredentials>) {
     super(data);
   }
@@ -31,4 +43,5 @@ export interface UserCredentialsRelations {
   // describe navigational properties here
 }
 
-export type UserCredentialsWithRelations = UserCredentials & UserCredentialsRelations;
+export type UserCredentialsWithRelations = UserCredentials &
+  UserCredentialsRelations;
